public_quiz/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-braces==1.15.0
sqlparse==0.4.3
